RD-6073 Compose: start all services (#3877)
* insert roles if they're missing

* configure_manager: set up logging

I'm gonna log more, so it's going to be useful

* configure_manager: cope with empty config

if there's no config at all, default a rabbitmq

* configure_manager: separate loading config

we want the file-based config to be loaded first, so that hash_salt is
already set (which happens in setup_flask_app).

The db-based config can be loaded later

* auth: allow the bootstrap admin to do everything

This is possibly to be reverted after we insert default roles/permissions

(or it can stay? It's probably a good idea anyway)

* configure_manager: add a way to create admin token

* configure_manager: add shorthands for waiting on services

* prepare_secrets.sh: include encryption_key

* when starting services, wait for db/rabbitmq

* add a script for generating the required certs

This is very temporary. We can't have hardcoded CN=nginx, of course

* run the mgmtworker

we can start it now!

* update db-schema.sql

* make config be just idempotent

no need to wait, if all the work can be safely re-done

* if cert is not found, just return empty

* tenant tests: mock the check function directly

now that we give admin all privileges always, just updating
permissions isn't quite going to cut it. But we can simply mock
the check function (and assert on the permission name too)