snapshot_utils: rewrite migrate_pickle_to_json (#3886)
* snapshot_utils: rewrite migrate_pickle_to_json

Rewrite using a loop rather than writing everything out.

Anyway, the real point of this change, is the `json_empty` check:
it's not always a json "'null'", because in case of inserting the
models directly via sql (eg. in snapshot-restore), it'll be just
a SQL null...

* migrations: allow plugins.wheels to be null

It is actually null in many (most? all?) snapshots

Co-authored-by: Mateusz Neumann <mateumann@gmail.com>