Fixup snapshot inte-tests (#3891)
* restore_idd_script: set current user in app context

It used to be set outside of app ctx, and so the user didnt actually
persist.

Also, no need to set those config paths. They're already defaulted
in config.py

* snapshot_utils: still copy the pickle if its falsy

this way we still copy over empty dicts

* inte-tests: load csv properly

Now that the values returned can be json and not pickle, they can
contain commas, so a split isn't enough anymore. We emit csv, we should
load with csv.

* snapshot tests: load using json

pickle's gone, man.

* flake