Use new StrEnum if available, else fallback
In 3.11, str()-ing an enum based off of `class x(str, Enum)` doesn't
give you the string value anymore - I'm not sure why... Anyway,
for 3.11-compat with a `class x(str, Enum)`, we can just define
a `__str__` explicitly.

And in 3.11, a new StrEnum is available, so let's use that.