RD-6290 Implement MonitoringAuth (#3909)
* allow errors to have additional headers

Now an error class will also be able to define headers to be returned

* return WWW-Authenticate when NoAuthProvided

Now users will be able to access some pages with their browser,
albeit with the ugly prompt.

Most of the api is still not usable though, because it requires
the tenant header as well.

But for /monitoring, this is going to be enough.

* RD-6290 Implement MonitoringAuth

When the user hits /monitoring, they need to pass through this
auth_request first.

Co-authored-by: geokala <geokala@users.noreply.github.com>