default value for Users.created_at
When creating a user, default their created_at to now.
Note: this also makes created_at be the default_sort_column.

Also, this causes a migration, because now, created_at is not nullable.
That means we must default the value to something, and well,
the something has to be the migration time.