RD-5952 Add a multi-instance IDD test
Well the test is pretty complex. But just read the docstring :)

Also, add a pre-check for a more readable error. This will happen
when you do `{get_capability: {get_attribute: ..}}` in case the
attribute returns a non-string. If we allowed the non-string to go
into the db select, we'd get a psycopg2 db error, and that one was
completely unreadable.