RD-5946 Reverse the "cyclic" check in create_idds
Reverse the check, from `if target in source.get_ancestors()`, to
`if source in target.get_dependencies()`.

I've reversed the check, but it COULD also be
`if target in source.get_dependents()`.

Point is, it should be dependents/dependencies, NOT as before,
ancestors/descendents.
This is because `ancestors` is DLD, and `dependents` is IDD. And here,
we really want to check for IDD.

It would be even better if it was `.get_all_dependents()` - both DLD
and IDD. Funnily, a single `PUT` create-idd DOES check that, but
the `POST` create-many uses this function that was changed here...
This is obviously not great, but I'm not gonna touch that here.

The problem that stops us from fixing this directly, is that sometimes
we DO want a "cyclic" dependency - an environment who creates a
Component, and makes the Component deployment be its child - then
the environment depends on the component via IDD (by way of being
a component), but the component deployment depends on the environment
via DLD (by way of csys-obj-parent).

So in the future hopefully we'll be able to change this (and so make
the `PUT` and `POST` ones use the same method). FWIW, deployment
creation uses the `POST` create_many, and Component uses the `PUT`.

So with this change, the following obvious blueprint will now work:
(before, it was considered "cyclic")
```yaml
labels:
    csys-obj-parent:
        values:
            - d1
outputs:
    out1:
        value: {get_capability: [d1, cap1]}
```

PS. The type declaration was wrong. `dependency` is the source dict,
not an IDD object.