`sm.transaction`: allow a way for the caller to control commit/rollback
If the transaction block throws an error, we normally would just
rollback, but what if we wanted to commit?

The use-case is when we do something in a transaction, it fails, and
then we want to set the relevant object's `.status` to "failed".
In that case, we do want to be able to commit and re-throw, and still
keep the status.