WDR-150 Don't lose the DepUpdate object on dependency errors
If a dependency is affected, and we'd like to mark the DepUpdate as
failed, we gotta do that in the transaction, and STILL COMMIT. If we
rollback, we leave the dep-update in the "updating" state.

Before, the situation was even funnier: we didn't commit (or even
flush), so the "find current DepUpdate" query would throw a 404,
that would rollback the transaction, and instead of a "dependency
affected" error or similar, you'd just get a "404 deployment update
not found". When starting a deployment update. Especially irritating
when starting many of them as part of an execution group.

Additionally, move the `prepare_executions` to _after_ the other stuff.
We want to render this as late as possible: in this case, so the
possibly-failed execution, would have a chance to get added to the
exec-group.

This is tested by the just-added `test_parent_child`, where we fail
with an "existing installations depened on..." error message, rather
than a 404.