evaluate workflow parameters: pass SM through
When refreshing an execution for dequeue, we'll be in context of
another tenant possibly (of the just-finished execution, that
caused us to dequeue);  and now that we evaluate all kinds of
parameters here (mostly wrt. the new object types), we do need to pass
in a SM that has the _correct_ tenant baked in - the execution's
tenant, not the current tenant.

I was only able to figure this out because I added those additional
prefixes to the error message, so let's keep those...

This is mostly unrelated to the rest of the commits here, but this
fixes `test_schedule_execution_snapshot_running_multi_tenant`