RD-5946 Fix the "cyclic dependency" check for dep-update (#3936)
* RD-5946 Reverse the "cyclic" check in create_idds

Reverse the check, from `if target in source.get_ancestors()`, to
`if source in target.get_dependencies()`.

I've reversed the check, but it COULD also be
`if target in source.get_dependents()`.

Point is, it should be dependents/dependencies, NOT as before,
ancestors/descendents.
This is because `ancestors` is DLD, and `dependents` is IDD. And here,
we really want to check for IDD.

It would be even better if it was `.get_all_dependents()` - both DLD
and IDD. Funnily, a single `PUT` create-idd DOES check that, but
the `POST` create-many uses this function that was changed here...
This is obviously not great, but I'm not gonna touch that here.

The problem that stops us from fixing this directly, is that sometimes
we DO want a "cyclic" dependency - an environment who creates a
Component, and makes the Component deployment be its child - then
the environment depends on the component via IDD (by way of being
a component), but the component deployment depends on the environment
via DLD (by way of csys-obj-parent).

So in the future hopefully we'll be able to change this (and so make
the `PUT` and `POST` ones use the same method). FWIW, deployment
creation uses the `POST` create_many, and Component uses the `PUT`.

So with this change, the following obvious blueprint will now work:
(before, it was considered "cyclic")
```yaml
labels:
    csys-obj-parent:
        values:
            - d1
outputs:
    out1:
        value: {get_capability: [d1, cap1]}
```

PS. The type declaration was wrong. `dependency` is the source dict,
not an IDD object.

* inte-tests: allow `self.deploy` to set labels too

Gonna be useful.

* Add dep-update inte-tests

* `sm.transaction`: allow a way for the caller to control commit/rollback

If the transaction block throws an error, we normally would just
rollback, but what if we wanted to commit?

The use-case is when we do something in a transaction, it fails, and
then we want to set the relevant object's `.status` to "failed".
In that case, we do want to be able to commit and re-throw, and still
keep the status.

* WDR-150 Don't lose the DepUpdate object on dependency errors

If a dependency is affected, and we'd like to mark the DepUpdate as
failed, we gotta do that in the transaction, and STILL COMMIT. If we
rollback, we leave the dep-update in the "updating" state.

Before, the situation was even funnier: we didn't commit (or even
flush), so the "find current DepUpdate" query would throw a 404,
that would rollback the transaction, and instead of a "dependency
affected" error or similar, you'd just get a "404 deployment update
not found". When starting a deployment update. Especially irritating
when starting many of them as part of an execution group.

Additionally, move the `prepare_executions` to _after_ the other stuff.
We want to render this as late as possible: in this case, so the
possibly-failed execution, would have a chance to get added to the
exec-group.

This is tested by the just-added `test_parent_child`, where we fail
with an "existing installations depened on..." error message, rather
than a 404.

* snap-restore: fixup update_deployment_dependencies_from_plan

- Actually call the normalize func. Whoops :)
- Never throw a "cyclic" error. If it was possible to be created before
  the snapshot, it should still be stored after the restore... that's
  not ideal, but the alternative is, what, just aborting the snapshot.
  Eh. Better give the user the broken data, which was broken before
  as well - they can still force-uninstall it...

* remove unused functions

how long have these been unused? since some rewrite probably. Months?
Well, free -lines.

* fixup migrations

Whoops - removing the constraints obviously must happen before removing
the table.
(maybe we could just not remove them, but the one on secrets we
probably need to keep...)

* evaluate workflow parameters: pass SM through

When refreshing an execution for dequeue, we'll be in context of
another tenant possibly (of the just-finished execution, that
caused us to dequeue);  and now that we evaluate all kinds of
parameters here (mostly wrt. the new object types), we do need to pass
in a SM that has the _correct_ tenant baked in - the execution's
tenant, not the current tenant.

I was only able to figure this out because I added those additional
prefixes to the error message, so let's keep those...

This is mostly unrelated to the rest of the commits here, but this
fixes `test_schedule_execution_snapshot_running_multi_tenant`

* Fix a few tests

- declare the new `reinstall` workflow. We've had it for a while, but
  it's time to have the test stop failing too
- replace `assertRegexpMatches` with just a `==`, to remove
  deprecation warnings
- just rewrite that permissions test already. It's been flaky for too
  long (mostly happens with xdist >1). This was because roles.first()
  wasn't always the admin role, but was sometimes an unrelated empty
  role that didn't have any permissions (the query doesn't have
  an order_by). So instead, let's make a Role right here, and then
  we'll always know what we're asserting on.