RD-6476 Rewrite upload_manager.py (#3939)
* Rewrite UploadedBlueprintsManager as a function

Instead of having a whole class'n'stuff, just have an upload_blueprint
function, which stores the blueprint files.
The endpoint is going to do the DB stuff.

* Blueprints: handle inserting the blueprint into the db

This is now going to happen on the endpoint side, rather than
on the upload_manager side. I prefer upload_manager to only handle
uploads, not actually creating db entries.

* Use a transaction for inserting blueprints

* Use an `.uploading` dir for uploading blueprints

Rather than store the in-flight received data in a tmpdir, store
them closer to their final destination, in an attempt to always use
the same filesystem (avoiding a copy)

* Replace UploadedBlueprintsValidator

Similar to uploading blueprints, validation now also does the DB stuff
in the endpoint.

* Rewrite uploading plugins & caravans

Similar to blueprints, upload_manager will now only store the files,
and the endpoint will now store the db resources.

upload_manager still needs to examine the uploaded plugin, and return
wagon_info (for one, or multiple, plugins)

* Rewrite UploadedSnapshotsManager

Similar to blueprints and plugins, snapshots are now going to use
a simple `upload_snapshot` function as well.

* Cleanup - remove unused stuff

* make sure the "uploading" temp folders are cleaned up

try/finally: rmtree for all the upload calls. We must not leak anything
even in face of errors

- _store_plugin can't really be a generator, because then the finally
  removes the directory before the whole generator gets executed...
- 755 is the right chmod, not 644 - this is a directory

* empty commit to kick CI

Co-authored-by: Mateusz Neumann <mateumann@gmail.com>
Co-authored-by: glukhman <47591609+glukhman@users.noreply.github.com>