fixup migrations
Whoops. My bad.

With this, migrations can again be rendered into a .sql that can be
loaded correctly.

After merging this, make sure to run build-rpms or upload-db-schema
on jenkins, so that the db schema in s3 is updated, and the postgres
container (in the next commit) can use it.