Separate services: simplify `docker compose up` (#3943)
* remove db-schema.yaml

We're not gonna need it anymore. I'll do something else in one of the
next commits, but i'll remove this file in a commit of its own,
because with this many -lines, it'd make other files unviewable.

* fixup migrations

Whoops. My bad.

With this, migrations can again be rendered into a .sql that can be
loaded correctly.

After merging this, make sure to run build-rpms or upload-db-schema
on jenkins, so that the db schema in s3 is updated, and the postgres
container (in the next commit) can use it.

* Use a postgres image with the schema baked in

Download the schema into the docker image.
That schema is built in jenkins, and uploaded to s3.

Sorry, but due to what I mentioned in the previous commit, it won't work
out of the box - you do need to run the build after merging this :)

* Simple development mode - allow non-ssl, default password

I want to be able to just `cfy profiles use`. Let's allow non-ssl,
and default the admin password to "admin", rather than random.

This is """temporary""", but I don't expect this to change before
phase 4 or 5.

* mgmtworker: separate entrypoint.sh

Allow waiting for the admin-token to be created, so that we don't
restart unnecessarily, creating a lot of noise in the log.

* restservice: download the license public key

To allow `cfy lic upl`, we do need the public key.

Are you surprised we just keep that key on github? I was too.

* prepare_secrets.sh: hide output

No need to show the `echo` in the log - it might (and does) contain
secrets/credentials!

* config: wait for db & rabbitmq

Rather than restarting a LOT, just wait for the components to be up.

Also, no need to keep restarting this container.
Once it's done, it's done.

* flake