RD-6293 Executions cascade: handle all node-instances (#3950)
* Add tests for cascading executions

I note that the whole cascade-execution infra doesn't have any tests
at all, in this suite!
I return [] from `_find_all_components_deployment_id`, and the whole
suite still passes... not great.

So I add some tests for it here.

(it was still tested in inte-tests, though)

* RD-6293 Executions cascade: handle all node-instances

- when there's no node-instances for a component node, don't break
- when there's >1 instances for a component node, handle all of them

* Rewrite the method using DB methods

This is going to be much faster! Instead of doing multiple queries
and filtering python-side, we can take advantage of the JSON operators
in postgres, now that we have JSON columns rather than pickle.