RD-6332 update: run check_drift tasks by dependecy order (#3949)
Users wanted it this way (in CYBL-1867), and I don't see why not.
I'd like to avoid adding a configuration flag too. This is too minor
to really give a switch. Might well just do it like this.

Also, since we're doing this in update, let's also set true to be
the default for the standalone check_drift, for a more consistent
user experience.

Co-authored-by: geokala <geokala@users.noreply.github.com>