Several test cleanups: drop server.reset_app (#3954)
* Remove server.reset_app

Turns out it's just not needed. Neither is the hash-mechanism-plaintext.
Copy resources is still needed though. Unfortunate.

* Move copy_resources to base_test.py

This is the only place it's used, so let's drop it from the actual
code, and only have it in tests.

Also simplify it a bit. It still very closely depends on the filesystem
but is more explicit now.

* tests: remove put_mock_deployments

Why was there a method for putting exactly two (not one, not three)
deployments... :) Replace it with a `._deployment()` on the local
class, not on the base class. More similar to all the other usage
we have.

* replace utils.mkdirs with os.makedirs

It was only ever used in this one place! And os.makedirs has just
a flag for this try/except behaviour anyway.

* remove duplicated get_amqp_client

We've had two implementations of get_amqp_client, differing by vhost
and by connect_timeout.
It can only be just one!
The status check still needs a connect timeout, because we want to
be able to return a failure in a finite amount of time.

* flake