agent+snapshot inte-tests: do the download/upload dance
Now that restore(force=True) doesn't reset storage by itself anymore,
we need to (same as we did in agentless tests) reset it ourselves,
and so, also need to upload the snapshot zip post-reset.

Also, I'm gonna change one of them to uninstall the agents AFTER the
restore, for additional coverage: now we also check that we can connect
to the agent after it's been restored.