Agents to_response: use get_data
Base including the rabbitmq credentials on get_data, rather than on
include. It's more similar to what we do in other places I think.

Also easier for the client side, because no need to declare all
the fields, if we just want to omit credentials (which is the
common case anyway)