RD-6537 inte-tests: fix agent_tests/test_snapshot.py (#3958)
* agent+snapshot inte-tests: do the download/upload dance

Now that restore(force=True) doesn't reset storage by itself anymore,
we need to (same as we did in agentless tests) reset it ourselves,
and so, also need to upload the snapshot zip post-reset.

Also, I'm gonna change one of them to uninstall the agents AFTER the
restore, for additional coverage: now we also check that we can connect
to the agent after it's been restored.

* agent snapshot inte-tests: concurrent deploy+undeploy

Those tests deploy 2 or 3 agents. No need for that be serial, we can
just start all executions first, and only then wait for all of them.

* agents.list() no longer has an implicit filter

This is the other part of cloudify-cosmo/cloudify-common#1198

* Agents to_response: use get_data

Base including the rabbitmq credentials on get_data, rather than on
include. It's more similar to what we do in other places I think.

Also easier for the client side, because no need to declare all
the fields, if we just want to omit credentials (which is the
common case anyway)