Rewrite component.check_drift
Instead of using dict_diff for those attributes, compare them
explicitly. I think that's better, because it allows us to easily
decide _how_ should they be compared

And indeed, in this, I'm also making sets out of labels, so that label
order doesn't matter.

Also, I'm going to ignore checks, if the items are missing...