RD-6635 Component check_drift: handle auto_inc_suffix (#3988)
* refactor component tests a bit

- base test: make .get_mock_ctx a bit more useful (now returns a
  ready-to-use ctx, and tests can call it)
- restclient mocks: return proper objects, not dicts
- test_blueprint_handling: rewrite a test to use .set_existing_objects
  rather than depending on blueprint/deployment IDs just happening to
  match
- test_blueprint_handling: drop duplicated test. That _success one
  did exactly the same thing as the one I've just rewritten.

* add basic check_drift tests

* Rewrite component.check_drift

Instead of using dict_diff for those attributes, compare them
explicitly. I think that's better, because it allows us to easily
decide _how_ should they be compared

And indeed, in this, I'm also making sets out of labels, so that label
order doesn't matter.

Also, I'm going to ignore checks, if the items are missing...

* RD-6635 add an auto_inc_suffix test for check_drift

Right now it fails :)

* RD-6635 check_drift: ignore suffix if auto_inc_suffix is set

* throw in an inte-test for good measure

and yes I checked that it fails when I remove the auto_inc_suffix bit

* Also handle blueprint labels drift correctly

Heh, it wasn't quite right before.