RD-6613 LocalStorageHandler.proxy: use `path` correctly
make_streaming_response expects to get a path under /resources, ie.
the URL part of a file to send. This is because X-Accel-Redirect wants
a location in nginx, not a path to a file.

So, from the restservice, we return X-Accel=/resources/foo.txt,
nginx goes to the /resources location, and sees the
`alias {{ manager.file_server_root }};` directive there (on AIO),
and serves file from that.

Also, no need to check for file existence right here. Nginx will
just do it.