RD-6588 6.4.2 backports (#4007)
* RD-6652 Add a test for scale rollback operations (#4006)

This is a test for cloudify-cosmo/cloudify-common#1219

* RD-6635 Component check_drift: handle auto_inc_suffix (#3988)

* refactor component tests a bit

- base test: make .get_mock_ctx a bit more useful (now returns a
  ready-to-use ctx, and tests can call it)
- restclient mocks: return proper objects, not dicts
- test_blueprint_handling: rewrite a test to use .set_existing_objects
  rather than depending on blueprint/deployment IDs just happening to
  match
- test_blueprint_handling: drop duplicated test. That _success one
  did exactly the same thing as the one I've just rewritten.

* add basic check_drift tests

* Rewrite component.check_drift

Instead of using dict_diff for those attributes, compare them
explicitly. I think that's better, because it allows us to easily
decide _how_ should they be compared

And indeed, in this, I'm also making sets out of labels, so that label
order doesn't matter.

Also, I'm going to ignore checks, if the items are missing...

* RD-6635 add an auto_inc_suffix test for check_drift

Right now it fails :)

* RD-6635 check_drift: ignore suffix if auto_inc_suffix is set

* throw in an inte-test for good measure

and yes I checked that it fails when I remove the auto_inc_suffix bit

* Also handle blueprint labels drift correctly

Heh, it wasn't quite right before.

* RD-6288 dep-update: bring back `reinstall_list` (#3982)

* RD-6288 dep-update: bring back `reinstall_list`

CLI says:
```
  -r, --reinstall-list TEXT      Node instances ids to be reinstalled as part
                                 of deployment update. They will be
                                 reinstalled even if the flag --skip-reinstall
                                 has been supplied
```

But that was somehow missing from the dep-update. Heh. Now it's back,
and now it has a test, too.

* flake

* RD-6633 Update: don't `establish` non-installed relationships (#3984)

* Make a test depicting the situation

* dep-update: when reinstalling node, only use started nodes as "intact"

* RD-6634 Update: run reinstalls before new installs (#3985)

* Add a test

* Update: run reinstalls before installs

* RD-6290 Implement MonitoringAuth (#3909)

* allow errors to have additional headers

Now an error class will also be able to define headers to be returned

* return WWW-Authenticate when NoAuthProvided

Now users will be able to access some pages with their browser,
albeit with the ugly prompt.

Most of the api is still not usable though, because it requires
the tenant header as well.

But for /monitoring, this is going to be enough.

* RD-6290 Implement MonitoringAuth

When the user hits /monitoring, they need to pass through this
auth_request first.

Co-authored-by: geokala <geokala@users.noreply.github.com>

* RD-6717 MonitoringAuth: use stage login page (#4004)

* Remove additional headers from the default 401

Not actually needed. We don't ever want the browser to auth here
I think.

* MonitoringAuth: use stage's login

This is a bit... dirty? But it's going to be nicer for the user, to
be able to reuse stage's login for this, rather than the browser popup

* RD-6293 Add tests for cascading executions

I note that the whole cascade-execution infra doesn't have any tests
at all, in this suite!
I return [] from `_find_all_components_deployment_id`, and the whole
suite still passes... not great.

So I add some tests for it here.

(it was still tested in inte-tests, though)

* RD-6293 Executions cascade: handle all node-instances

- when there's no node-instances for a component node, don't break
- when there's >1 instances for a component node, handle all of them

* python 3.6-compatible syntax

* handle None runtime_props

* set() the deployment ids

Similarly on master, we do have .distinct() there

* RD-6284 types.yaml: new workflow parameters (#3910)

* RD-6284 types.yaml: new workflow parameters

- Add filtering parameters to `install`
- Add filtering parameters to `uninstall`
- Declare the `reinstall` workflow

This is done in all types (1_5, 1_4, 1_3), because the parameters
don't depend on dsl version (only on the manager version).
However 1_3 of course doesn't get 1_4+ dsl features,
ie. availability_rules and item_type

* test_deployments: there's now 16 workflows

after adding `reinstall`

* add missing test blueprints

Co-authored-by: geokala <geokala@users.noreply.github.com>