RD-6766 mgmtworker dockerfile: install the missing libs (#4048)
* mgmtworker dockerfile: install the missing libs

Those are actually required to install reqs.txt - I think it was missed
before due to all the caching I had going on, but now that I actually
ran that clean, I noted that these are needed.

* restservice: allow building w/o premium

For the life of me, I can currently figure out how to give jenkins
the ssh credentials for cloudify-premium. Well. It's probably a good
idea to parametrize this anyway.