RD-6658 inte-tests: update expected plugin path
After cloudify-cosmo/cloudify-agent#867 the plugins are put in the
agent dir, rather than near the env.

Anyway, the env is usually going to be like /opt/cloudify-agent-7.0.0/
and the agent user won't necessarily going to have write privileges
for that dir (indeed, it had better not have write rights there).

So the agent must store the plugins in a directory where it does
have write rights, like right here.

(in the case of this test, basedir and agent dir just happen to be
the same...)