RD-7052 configure_manager: make create-admin-token separate
The "create-admin-token" part will now be a "subcommand" similar to
db-wait and rabbitmq-wait.

The reason is, we want to be able to run create-admin-token, without
running the actual "configure" parts.

And that is because normally create-admin-token runs in the mgmtworker
initcontainer, while the configure parts run in the restservice one.
So that's why we want them to be separate.