RD-7061 dep-update inte-tests: use `get` in scripts
All the scripts used something like:
```bash
a_ops_counter=$(ctx instance runtime_properties 'a_ops_counter')
```
On the first run of the script, that errors out, spewing an error
message into the logs - because that is a KeyError, the property
doesn't exist.

The script continues anyway, because there's no `set -e`, and the test
can pass.

This commit replaces it with a `get` in all the scripts, so that
there's no unnecessary error printout.

(and now the scripts COULD use `set -e` if we wanted to)