RD-206 Replace sm.full_access_list with direct queries (#4103)
Turns out we don't really need .full_access_list: it's used in only
a few places, and it's easily replaceable by just direct queries.
I prefer that, and it'll make the upcoming changes easier, by reducing
the storage-manager scope somewhat.

Also, for the maintenance-mode test: I replace the python-side forloop
with a query - so I also change the test to not have status=None,
but a valid status, because I don't really want to make the query
have to cover None.  (and if an execution does have None somehow - we
don't really want to block maintenance on it anyway. So that used to
be a bug, really)