Rewrite storage-manager query building
This essentially rewrites `_get_query`. The new implementation
is a bit shorter, because it attempts to do all the query modifications
inside of _get_query itself, and the other sub-methods only generate
clauses to filter/sort on.

So, we resolve fields only once (including assoc-proxies) and then
apply filters on those fields.

The main reason for this was to make assoc-proxy resolving actually
correct, which I think it now is, with the aliased-outerjoin-getattr
dance in `_resolve_included_fields`. But after that was done, I found
that pretty much the whole method had to be rewritten to use those
resolved fields.