remove now-pointless test
You'll note that the new storage-manager _get_query (and resolve-fields)
doesn't have a `.load_only`. So even if `include` is limited, we still
load all default attributes.

This is because it's just not worth it. We'll still need a lot of
fields that aren't declared (all the FKs), so we're just loading
all the fields always. In performance testing there's essentially
no difference.

Note: those are only fields local to the row, not relationships. So
if we're already looking up the row, might as well grab the whole
thing.

We might want to go back on this, eventually, but we'd need to figure
out a way to also include required FKs as well. For right now, this
is acceptable I think.