RND-206 storage_manager._get_query - load correct assoc-proxies (#4109)
* Rewrite storage-manager query building

This essentially rewrites `_get_query`. The new implementation
is a bit shorter, because it attempts to do all the query modifications
inside of _get_query itself, and the other sub-methods only generate
clauses to filter/sort on.

So, we resolve fields only once (including assoc-proxies) and then
apply filters on those fields.

The main reason for this was to make assoc-proxy resolving actually
correct, which I think it now is, with the aliased-outerjoin-getattr
dance in `_resolve_included_fields`. But after that was done, I found
that pretty much the whole method had to be rewritten to use those
resolved fields.

* don't pass `load_relationships`

This argument is now gone, and it's not needed, because now we're
automatically doing the outerjoins anyway. And so, the property is
unnecessary as well.

* add an IDD filtering test

this is pretty much what I've set out to make pass, in this PR.
Now, filtering by assoc-proxies works correctly, even when there's two
proxies to the same table.

* agents endpoint: rename the filter

the assoc-proxy is named _id, not _ids. But for back-compat, let's
still allow the old param name as well.

* sites: replace .exists with .get

This was pretty much the only use of .exists, and so, .exists was
removed from storage-manager.

* deployment render: use latest_execution_id in fields

the assoc-proxy is named latest_execution_id, even if we rename it
in the actual json rendering

* deployments render: don't include schedules

It's VERY expensive to include those, and they're never used. Not even
exposed on the restclient, and not used in the UI.
Let's just not.

* remove now-pointless test

You'll note that the new storage-manager _get_query (and resolve-fields)
doesn't have a `.load_only`. So even if `include` is limited, we still
load all default attributes.

This is because it's just not worth it. We'll still need a lot of
fields that aren't declared (all the FKs), so we're just loading
all the fields always. In performance testing there's essentially
no difference.

Note: those are only fields local to the row, not relationships. So
if we're already looking up the row, might as well grab the whole
thing.

We might want to go back on this, eventually, but we'd need to figure
out a way to also include required FKs as well. For right now, this
is acceptable I think.

* simplify one_to_many_relationship

`foreign_keys` is a simpler way of writing that primaryjoin. Sqlalchemy
does this automatically.

* flake

* s/if not/is not None/