RD-7094 Usable audit_log reference to the original object (#4118)
* Implement new write_audit_log triggered function

The function is an enhancement from the previously used
write_audit_log_storage_id and write_audit_log_id functions.  Now it
allows to store arbitrary unique identifier fields for the referenced
object.  E.g. for blueprints we are going to store _tenant_id and id,
roles - just an id etc.  These values will enable us to retrieve
referenced objects more easily.

* Do not enforce not-nullable plugins.wheels on downgrade

Alembic downgrade to Cloudify 6.4 should not enforce plugins.wheels_p to
be not nullable.  That field is practically not used in Cloudify 7.0
(hence it will most likely be empty).  And also we do not intend to
support downgrade to older Cloudify versions.

* Update resource_model

* Iterate over id-columns should improve performance

Co-authored-by: Łukasz Maksymczuk <tehasdf@users.noreply.github.com>

* Comments

---------

Co-authored-by: Łukasz Maksymczuk <tehasdf@users.noreply.github.com>