RND-328 execution-scheduler: update config when it changes
Commonly, the scheduler starts before the restservice does, and in
the case of the k8s deployment, even before the restservice is
finished inserting rabbitmq address & credentials into the db.

Then, the scheduler fails to run anything, because it started and
queried the config before the rabbitmq address was populated there.

Restarting the scheduler would mitigate that, but it's of course
not practical to restart the scheduler after every change to the
config (including the initial "change" - first inserting it).

So, let's use the same method we use in the restservice, and query
last-changed times for the configs. To do that, move the function
to the utils module, so that it can be imported into the scheduler
service (importing server.py has side-effects, unfortunately...)